+++ fastapi_opa/opa/opa_config.py
@@ -1,3 +1,4 @@
+import re
 from abc import ABC
 from abc import abstractmethod
 from typing import List
@@ -9,9 +10,11 @@ from fastapi_opa.auth.auth_interface import AuthInterface
 
 
 class Injectable(ABC):
-    def __init__(self, key: str, skip_endpoints: list = []) -> None:
+    def __init__(
+        self, key: str, skip_endpoints: Optional[List[str]] = []
+    ) -> None:
         self.key = key
-        self.skip_endpoints = skip_endpoints
+        self.skip_endpoints = [re.compile(skip) for skip in skip_endpoints]
 
     @abstractmethod
     async def extract(self, request: Request) -> List:
@@ -24,7 +27,9 @@ class OPAConfig:
         authentication: AuthInterface,
         opa_host: str,
         injectables: Optional[List[Injectable]] = None,
+        accepted_methods: Optional[List[str]] = ["id_token", "access_token"],
     ) -> None:
         self.authentication = authentication
         self.opa_url = f"{opa_host.rstrip('/')}/v1/data/httpapi/authz"
         self.injectables = injectables
+        self.accepted_methods = accepted_methods